fix: remove engine dependency
