+++ cortex-js/src/infrastructure/commanders/engines/engines-init.command.ts
@@ -4,8 +4,8 @@ import { ContextService } from '@/infrastructure/services/context/context.servic
 import { Engines } from '../types/engine.interface';
 import { CortexUsecases } from '@/usecases/cortex/cortex.usecases';
 import { FileManagerService } from '@/infrastructure/services/file-manager/file-manager.service';
-import { EnginesUsecases } from '@/usecases/engines/engines.usecase';
 import { BaseCommand } from '../base.command';
+import { defaultInstallationOptions } from '@/utils/init';
 
 @SubCommand({
   name: '<name> init',
@@ -17,7 +17,6 @@ import { BaseCommand } from '../base.command';
 @SetCommandContext()
 export class EnginesInitCommand extends BaseCommand {
   constructor(
-    private readonly engineUsecases: EnginesUsecases,
     private readonly cortexUsecases: CortexUsecases,
     private readonly fileManagerService: FileManagerService,
     readonly contextService: ContextService,
@@ -32,7 +31,7 @@ export class EnginesInitCommand extends BaseCommand {
     const engine = passedParams[0];
     const params = passedParams.includes(Engines.llamaCPP)
       ? {
-          ...(await this.engineUsecases.defaultInstallationOptions()),
+          ...(await defaultInstallationOptions()),
           ...options,
         }
       : {};
